+++ service-worker.js
@@ -10,6 +10,7 @@ const ASSETS = [
   './css/tokens.css',
   './js/core/app.js',
   './js/platform/shell/themeController.js',
+  './js/platform/shell/settingsController.js',
   './js/core/centralStore.js',
   './js/core/domUpdate.js',
   './js/core/eventDelegation.js',